+++ CHANGELOG.md
@@ -1,6 +1,7 @@
 # v1.16.1 - Torrhus Canyon Patches
 
 ### Features
+- Extended the survivalist achievement with 4 new stages
 
 ### Fixes
 - Fixed other messages being counted as trophy catches when they shouldn't